feat: update api

+++ larksuite-oapi/src/main/java/com/lark/oapi/service/ehr/v1/model/SystemFields.java
@@ -375,13 +375,13 @@ public class SystemFields {
      * <p> 示例值：1608690517811
      */
     @SerializedName("create_time")
-    private Long createTime;
+    private Integer createTime;
     /**
      * 更新时间
      * <p> 示例值：1608690517811
      */
     @SerializedName("update_time")
-    private Long updateTime;
+    private Integer updateTime;
 
     // builder 开始
     public SystemFields() {
@@ -1171,19 +1171,19 @@ public class SystemFields {
         this.employeeFormStatus = employeeFormStatus;
     }
 
-    public Long getCreateTime() {
+    public Integer getCreateTime() {
         return this.createTime;
     }
 
-    public void setCreateTime(Long createTime) {
+    public void setCreateTime(Integer createTime) {
         this.createTime = createTime;
     }
 
-    public Long getUpdateTime() {
+    public Integer getUpdateTime() {
         return this.updateTime;
     }
 
-    public void setUpdateTime(Long updateTime) {
+    public void setUpdateTime(Integer updateTime) {
         this.updateTime = updateTime;
     }
 
@@ -1487,12 +1487,12 @@ public class SystemFields {
          * 创建时间
          * <p> 示例值：1608690517811
          */
-        private Long createTime;
+        private Integer createTime;
         /**
          * 更新时间
          * <p> 示例值：1608690517811
          */
-        private Long updateTime;
+        private Integer updateTime;
 
         /**
          * 中文姓名
@@ -2448,7 +2448,7 @@ public class SystemFields {
          * @param createTime
          * @return
          */
-        public Builder createTime(Long createTime) {
+        public Builder createTime(Integer createTime) {
             this.createTime = createTime;
             return this;
         }
@@ -2461,7 +2461,7 @@ public class SystemFields {
          * @param updateTime
          * @return
          */
-        public Builder updateTime(Long updateTime) {
+        public Builder updateTime(Integer updateTime) {
             this.updateTime = updateTime;
             return this;
         }
